requirements.txt
django~=5.1.0
channels[daphne]
django-cors-headers==4.4.0
django-cotton==0.9.40
django-environ==0.11.2
openai==1.45.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
crispy-tailwind==1.0.3

faker==28.4.1
factory-boy==3.3.1
